+++ packages/grid/src/features/standard-table/config/StandardTableConfig.ts
@@ -157,6 +157,11 @@ export interface StandardTableConfig<
   stickyHeader?: boolean;
 
   /**
+   * Enable sticky behaviour for header and row checkbox
+   */
+  stickyCheckboxColumn?: boolean;
+
+  /**
    * Set a custom z index
    */
   zIndex?: number;