+++ packages/grid/src/features/standard-table/features/checkboxes/StandardTableRowCheckbox.tsx
@@ -3,7 +3,10 @@ import { Checkbox, CheckboxProps } from "@stenajs-webui/forms";
 import * as React from "react";
 import { useGridCell } from "../../../grid-cell/hooks/UseGridCell";
 import { useTotalNumColumnsForRows } from "../../context/GroupConfigsForRowsContext";
-import { useStandardTableId } from "../../hooks/UseStandardTableConfig";
+import {
+  useStandardTableConfig,
+  useStandardTableId,
+} from "../../hooks/UseStandardTableConfig";
 
 interface Props extends Pick<CheckboxProps, "value" | "onValueChange"> {
   colIndex: number;
@@ -24,6 +27,7 @@ export const StandardTableRowCheckbox: React.FC<Props> = React.memo(
     const totalNumColumns = useTotalNumColumnsForRows();
 
     const tableId = useStandardTableId();
+    const { zIndex, stickyCheckboxColumn } = useStandardTableConfig();
     const gridCell = useGridCell<boolean>(Boolean(value), {
       colIndex,
       numCols: totalNumColumns,
@@ -40,6 +44,10 @@ export const StandardTableRowCheckbox: React.FC<Props> = React.memo(
         width={"45px"}
         minWidth={"45px"}
         indent
+        background={stickyCheckboxColumn ? "inherit" : undefined}
+        position={stickyCheckboxColumn ? "sticky" : undefined}
+        left={stickyCheckboxColumn ? "0px" : undefined}
+        zIndex={zIndex}
       >
         <Checkbox
           size={"small"}