+++ packages/select/src/components/ui/MultiSelect.tsx
@@ -4,23 +4,23 @@ import SelectComponent, { mergeStyles, Props } from "react-select";
 import { defaultSelectTheme, selectThemeDark } from "../../SelectTheme";
 import { createStylesFromTheme } from "../../util/StylesBuilder";
 
-export interface SelectProps<T = { label: string; value: string }>
-  extends Omit<Props<T, false>, "isMulti"> {
+export interface MultiSelectProps<T = { label: string; value: string }>
+  extends Omit<Props<T, true>, "isMulti"> {
   variant?: "dark" | "light";
 }
 
-export const Select = <T extends {}>({
+export const MultiSelect = <T extends {}>({
   variant = "light",
   styles,
   ...selectProps
-}: SelectProps<T>) => {
+}: MultiSelectProps<T>) => {
   const selectStyles = useMemo(() => {
-    const sourceStyles = createStylesFromTheme<T, false>(
+    const sourceStyles = createStylesFromTheme<T, true>(
       variant === "light" ? defaultSelectTheme : selectThemeDark
     );
 
     return styles ? mergeStyles(sourceStyles, styles) : sourceStyles;
   }, [variant, styles]);
 
-  return <SelectComponent styles={selectStyles} {...selectProps} />;
+  return <SelectComponent styles={selectStyles} isMulti {...selectProps} />;
 };